+++ src/helpers/casts.ts
@@ -1,20 +1,42 @@
-import {
-  InvoiceCurrency, InvoiceStatus, PaidBtnName, nonosToCoins,
-} from './utils';
+import { InvoiceCurrency, PaidBtnName, nonosToCoins } from './utils';
 
 /** Result type for {@link Client.getBalances} method */
 export type Balances = { [key: string]: string };
 
-export type CurrencyCode =
-  'USDT' | 'TON' | 'BTC' | 'LTC' | 'ETH' | 'BNB' | 'TRX' | 'USDC' | 'JET' | 'RUB' | 'USD' | 'EUR' |
-  'BYN' | 'UAH' | 'GBP' | 'CNY' | 'KZT' | 'UZS' | 'GEL' | 'TRY' | 'AMD' | 'THB' | 'INR' | 'BRL' |
-  'IDR' | 'AZN' | 'AED' | 'PLN' | 'ILS' | 'KGS' | 'TJS';
+// Crypto currency codes
+export type CryptoCurrencyCode =
+  'USDT' | 'TON' | 'BTC' | 'ETH' | 'LTC' | 'BNB' | 'TRX' | 'USDC' | 'JET';
 
-/** Possible currencies types */
+// Fiat currency codes
+export type FiatCurrencyCode =
+  'USD' | 'EUR' | 'RUB' | 'BYN' | 'UAH' | 'GBP' | 'CNY' | 'KZT' | 'UZS' | 'GEL' | 'TRY' | 'AMD' |
+  'THB' | 'INR' | 'BRL' | 'IDR' | 'AZN' | 'AED' | 'PLN' | 'ILS';
+
+// All curerencies codes, `getCurrencies` API method return also 'KGS' and 'TJS' fiat currencies,
+// so I add it's here
+// TODO: check availability of creating invoices with this currencies
+export type CurrencyCode = CryptoCurrencyCode | FiatCurrencyCode | 'KGS' | 'TJS';
+
+/** Possible currency types */
 export enum CurrencyType {
-  Blockchain = 'blockchain',
+  Crypto = 'crypto',
   Fiat = 'fiat',
+  Unknown = 'unknown',
+}
+
+/** Possible detailed currency types */
+export enum DetailedCurrencyType {
+  Blockchain = 'blockchain',
   Stablecoin = 'stablecoin',
+  Fiat = CurrencyType.Fiat,
+  Unknown = CurrencyType.Unknown,
+}
+
+/** Possible invoice statuses */
+export enum InvoiceStatus {
+  Active = 'active',
+  Paid = 'paid',
+  Expired = 'expired',
   Unknown = 'unknown',
 }
 
@@ -32,7 +54,7 @@ export type Currency = {
   /** Currency decimals count */
   decimals: number,
   /** Currency type */
-  type: CurrencyType,
+  type: DetailedCurrencyType,
 };
 
 /** Result type for {@link Store.getCurrencies} method */
@@ -69,22 +91,32 @@ export type Invoice = {
   status: InvoiceStatus,
   /** Invoice hash */
   hash: string,
+  /** Invoice currency type */
+  currencyType: CurrencyType,
   /** Invoice currency code */
   currency: InvoiceCurrency,
   /** Invoice amount */
   amount: number,
-  /** Invoice pay url for user */
-  payUrl: string,
+  /** Invoice pay url for user by bot */
+  botPayUrl: string,
+  /** Invoice pay url for user by mini app */
+  miniAppPayUrl: string,
+  /** Invoice pay url for user by web app */
+  webAppPayUrl: string,
   /** Is invoice allow user comment */
   isAllowComments: boolean,
   /** Is user can pay invoice anonymously */
   isAllowAnonymous: boolean,
   /** Invoice created date */
   createdAt: Date,
+  /** Text of the hidden message, only if set in invoice creation */
+  hiddenMessage?: string,
   /** Is invoice paid anonymously, only for paid invoice */
   isPaidAnonymously?: boolean,
   /** Invoice paid date, only for paid invoice */
   paidAt?: Date,
+  /** Expiration date, only if set pay limit time in invoice creation */
+  expirationDate?: Date,
   /** Invoice displayed to user description, only if `description` passed in invoice creation */
   description?: string,
   /**
@@ -108,6 +140,37 @@ export type Invoice = {
    * only if `paidBtnUrl` passed in invoice creation
    */
   paidBtnUrl?: string,
+  /**
+   * Asset of service fees charged when the invoice was paid, only if status is InvoiceStatus.Paid
+   */
+  feeAsset?: CryptoCurrencyCode,
+  /**
+   * Amount of service fees charged when the invoice was paid, only if status is InvoiceStatus.Paid
+   */
+  fee?: number,
+  /**
+   * Price of the asset in USD, only if status is InvoiceStatus.Paid
+   */
+  usdRate?: number,
+  /**
+   * List of assets which can be used to pay the invoice, only if set in invoice creation
+   */
+  acceptedAssets?: CryptoCurrencyCode[],
+  /**
+   * Cryptocurrency alphabetic code for which the invoice was paid,
+   * only if currency type is CurrencyType.Fiat and status is InvoiceStatus.Paid
+   */
+  paidAsset?: CryptoCurrencyCode,
+  /**
+   * Amount of the invoice for which the invoice was paid,
+   * only if currency type is CurrencyType.Fiat and status is InvoiceStatus.Paid
+   */
+  paidAmount?: number,
+  /**
+   * The rate of the paid_asset valued in the fiat currency,
+   * only if currency type is CurrencyType.Fiat and status is InvoiceStatus.Paid
+   */
+  paidFiatRate?: number,
 };
 
 /**
@@ -188,10 +251,10 @@ export const toCurrencies = (input: any): Currencies => {
     if (value.code) {
       const code: CurrencyCode = value.code.toString();
 
-      let type = CurrencyType.Unknown;
-      if (value.is_blockchain) type = CurrencyType.Blockchain;
-      if (value.is_fiat) type = CurrencyType.Fiat;
-      if (value.is_stablecoin) type = CurrencyType.Stablecoin;
+      let type = DetailedCurrencyType.Unknown;
+      if (value.is_blockchain) type = DetailedCurrencyType.Blockchain;
+      if (value.is_fiat) type = DetailedCurrencyType.Fiat;
+      if (value.is_stablecoin) type = DetailedCurrencyType.Stablecoin;
 
       const currency: Currency = {
         code: code,
@@ -238,22 +301,41 @@ export const toExchangeRates = (input: any): ExchangeRates => {
 export const toInvoice = (input: any): Invoice => {
   const invoice: Invoice = {
     id: input.invoice_id || 0,
-    status: input.status || '',
+    status: input.status || InvoiceStatus.Unknown,
     hash: input.hash || '',
-    currency: input.asset || '',
+    currencyType: input.currency_type || '',
+    currency: input.asset || input.fiat || '',
     amount: parseFloat(input.amount) || 0,
-    payUrl: input.pay_url || '',
     isAllowComments: input.allow_comments || false,
     isAllowAnonymous: input.allow_anonymous || false,
     createdAt: new Date(input.created_at),
+    botPayUrl: input.bot_invoice_url || '',
+    miniAppPayUrl: input.mini_app_invoice_url || '',
+    webAppPayUrl: input.web_app_invoice_url || '',
   };
 
+  if (invoice.currencyType === CurrencyType.Crypto) {
+    invoice.currency = input.asset || '';
+  }
+  if (invoice.currencyType === CurrencyType.Fiat) {
+    invoice.currency = input.fiat || '';
+  }
+
+  if (input.hidden_message !== undefined) invoice.hiddenMessage = input.hidden_message;
   if (input.paid_anonymously !== undefined) invoice.isPaidAnonymously = input.paid_anonymously;
+  if (input.expiration_date !== undefined) invoice.expirationDate = new Date(input.expiration_date);
   if (input.paid_at !== undefined) invoice.paidAt = new Date(input.paid_at);
   if (input.description !== undefined) invoice.description = input.description;
   if (input.paid_btn_name !== undefined) invoice.paidBtnName = input.paid_btn_name;
   if (input.paid_btn_url !== undefined) invoice.paidBtnUrl = input.paid_btn_url;
   if (input.comment !== undefined) invoice.comment = input.comment;
+  if (input.paid_usd_rate !== undefined) invoice.usdRate = parseFloat(input.paid_usd_rate) || 0;
+  if (input.fee_asset !== undefined) invoice.feeAsset = input.fee_asset || '';
+  if (input.fee_amount !== undefined) invoice.fee = input.fee_amount || 0;
+  if (input.accepted_assets !== undefined) invoice.acceptedAssets = input.accepted_assets;
+  if (input.paid_asset !== undefined) invoice.paidAsset = input.paid_asset || '';
+  if (input.paid_amount !== undefined) invoice.paidAmount = parseFloat(input.paid_amount) || 0;
+  if (input.paid_fiat_rate !== undefined) invoice.paidFiatRate = parseFloat(input.paid_fiat_rate) || 0;
   if (input.payload !== undefined) {
     let payload: any;